requirements.txt
flask==2.3.3
gunicorn==21.2.0
tensorflow==2.15.0
Pillow==10.2.0
flask-cors==4.0.0
numpy==1.24.3
keras==2.15.0
werkzeug==2.3.7